:recycle: [#1068] -- move data normalization to component plugins
