+++ src/openforms/formio/components/vanilla.py
@@ -4,6 +4,8 @@ Implement backend functionality for core Formio (built-in) component types.
 Custom component types (defined by us or third parties) need to be organized in the
 adjacent custom.py module.
 """
+from openforms.utils.date import format_date_value
+
 from ..formatters.formio import (
     CheckboxFormatter,
     CurrencyFormatter,
@@ -23,6 +25,7 @@ from ..formatters.formio import (
     TimeFormatter,
 )
 from ..registry import BasePlugin, register
+from ..typing import Component
 
 
 @register("default")
@@ -47,6 +50,10 @@ class Email(BasePlugin):
 @register("date")
 class Date(BasePlugin):
     formatter = DateFormatter
+
+    @staticmethod
+    def normalizer(component: Component, value: str) -> str:
+        return format_date_value(value)
 
 
 @register("time")
